requirements.txt
flask
ultralytics
opencv-python
paddleocr
numpy
Pillow
PyYAML
tqdm
requests
